+++ src/objects.js
@@ -94,7 +94,7 @@ embedMetadataPNG, SnapExtensions, SnapSerializer*/
 
 /*jshint esversion: 11*/
 
-modules.objects = '2023-April-25';
+modules.objects = '2023-April-28';
 
 var SpriteMorph;
 var StageMorph;
@@ -2476,7 +2476,17 @@ SpriteMorph.prototype.blockForSelector = function (selector, setDefaults) {
     if (contains(['reifyReporter', 'reifyPredicate'], block.selector)) {
         block.isStatic = true;
     }
+
     block.setSpec(localize(info.spec));
+
+    // localize the block. Replace the above line with the following one to
+    // enable translations for block specs to use padded underscores as
+    // placeholders for input slots.
+    // Under construction. Commented out for now (jens, 04/2023)
+    /*
+        block.setSpec(block.localizeBlockSpec(info.spec));
+    */
+
     if (migration && migration.expand) {
         block.inputs()[migration.expand].addInput();
     }